test: 添加 HashUtil 和 TimeUtil 的单元测试
- 新增 HashUtilTest 类，实现对 HashUtil 类的单元测试
- 新增 TimeUtilTest 类，实现对 TimeUtil 类的单元测试
- 修改 RandomMediaContentGenerator 类，将生成随机 ID 的方法从 int 改为 long

+++ src/test/java/com/memes/util/RandomMediaContentGenerator.java
@@ -33,7 +33,7 @@ public class RandomMediaContentGenerator {
     private static MediaContent generateRandomMediaContent() {
         return MediaContent
             .builder()
-            .id(random.nextInt())
+            .id(random.nextLong())
             .dataType(DataType.values()[random.nextInt(DataType.values().length)])
             .dataContent(dataContents[random.nextInt(dataContents.length)])
             .userId(userIds[random.nextInt(userIds.length)])
